source_code/requirements.txt
click==7.1.2
Flask==1.1.2
greenlet==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pip==20.2.3
psycopg2==2.8.6
python-dotenv==0.17.1
setuptools==49.2.1
SQLAlchemy==1.4.15
Werkzeug==1.0.1


